[BALL] Separate SIP- and PyInterpreter-related headers
