+++ include/BALL/PYTHON/EXTENSIONS/sources.cmake
@@ -1,8 +1,6 @@
 ### list all filenames of the directory here ###
-SET(GROUP PYTHON)
+SET(GROUP PYTHON/EXTENSIONS)
 
 FILE(GLOB HEADERS_LIST "include/BALL/${GROUP}/*.h" "include/BALL/${GROUP}/*.iC")
 
 ADD_BALL_HEADERS("${GROUP}" "${HEADERS_LIST}")
-
-INCLUDE(include/BALL/PYTHON/EXTENSIONS/sources.cmake)
