+++ source/PYTHON/EXTENSIONS/BALL/assignBondOrderProcessor.sip
@@ -2,7 +2,7 @@ class AssignBondOrderProcessor
 	: AtomContainerProcessor
 {
 %TypeHeaderCode
-	#include <BALL/PYTHON/pyUnaryProcessors.h>
+	#include <BALL/PYTHON/EXTENSIONS/pyUnaryProcessors.h>
 	#include <BALL/STRUCTURE/assignBondOrderProcessor.h>
 %End
