+++ source/PYTHON/EXTENSIONS/BALL/pyIndexList.C
@@ -1,10 +1,4 @@
-// -*- Mode: C++; tab-width: 2; -*-
-// vi: set ts=2:
-//
-// $Id: pyIndexList.C,v 1.2 2005-12-23 17:02:56 amoll Exp $
-//
-
-#include <BALL/PYTHON/pyIndexList.h>
+#include <BALL/PYTHON/EXTENSIONS/pyIndexList.h>
 #include <BALL/KERNEL/atom.h>
 #include <BALL/KERNEL/expression.h>
 #include <BALL/KERNEL/atomContainer.h>
@@ -19,7 +13,7 @@ namespace BALL
 
 	PyIndexList::PyIndexList(const PyIndexList& new_list)
 		: list<Index>(new_list)
-  {
+	{
 	}
 
 	PyIndexList::~PyIndexList()
@@ -27,7 +21,7 @@ namespace BALL
 	{
 	}
 
- 	PyIndexList::PyIndexList(const std::list<Index>& idx_list)
+	PyIndexList::PyIndexList(const std::list<Index>& idx_list)
 	{
 		operator = (idx_list);
 	}
@@ -37,7 +31,7 @@ namespace BALL
 		operator = (pos_list);
 	}
 
- 	PyIndexList::PyIndexList(const std::vector<Index>& idx_vector)
+	PyIndexList::PyIndexList(const std::vector<Index>& idx_vector)
 	{
 		operator = (idx_vector);
 	}
@@ -51,8 +45,8 @@ namespace BALL
 	{
 		// clear the old contents
 		clear();
-		
-		std::copy(idx_list.begin(), idx_list.end(), std::back_inserter<PyIndexList>(*this));		
+
+		std::copy(idx_list.begin(), idx_list.end(), std::back_inserter<PyIndexList>(*this));
 
 		return *this;
 	}
@@ -61,8 +55,8 @@ namespace BALL
 	{
 		// clear the old contents
 		clear();
-		
-		std::copy(pos_list.begin(), pos_list.end(), std::back_inserter<PyIndexList>(*this));		
+
+		std::copy(pos_list.begin(), pos_list.end(), std::back_inserter<PyIndexList>(*this));
 
 		return *this;
 	}
@@ -71,8 +65,8 @@ namespace BALL
 	{
 		// clear the old contents
 		clear();
-		
-		std::copy(idx_vector.begin(), idx_vector.end(), std::back_inserter<PyIndexList>(*this));		
+
+		std::copy(idx_vector.begin(), idx_vector.end(), std::back_inserter<PyIndexList>(*this));
 
 		return *this;
 	}
@@ -81,8 +75,8 @@ namespace BALL
 	{
 		// clear the old contents
 		clear();
-		
-		std::copy(pos_vector.begin(), pos_vector.end(), std::back_inserter<PyIndexList>(*this));		
+
+		std::copy(pos_vector.begin(), pos_vector.end(), std::back_inserter<PyIndexList>(*this));
 
 		return *this;
 	}
