+++ source/PYTHON/EXTENSIONS/BALL/ruleProcessor.sip
@@ -2,7 +2,7 @@ class RuleProcessor
 	: AtomProcessor
 {
 %TypeHeaderCode
-	#include <BALL/PYTHON/pyUnaryProcessors.h>
+	#include <BALL/PYTHON/EXTENSIONS/pyUnaryProcessors.h>
 	#include <BALL/MOLMEC/COMMON/ruleProcessor.h>
 %End
 	public:
